Rename zstd.rs to zstd_seekable_wrapper.rs
